docker_compose/simple_project/app/requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
dacite==1.6.0
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-extensions==3.2.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==3.20.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
split-settings==1.0.0
sqlparse==0.4.2
tzdata==2022.2
uWSGI==2.0.20
zipp==3.8.1